Update version date and made minor changes to the Response Time report

+++ beans/src/main/java/org/sola/clients/beans/application/ResponseBean.java
@@ -46,7 +46,7 @@ public class ResponseBean extends AbstractIdBean {
         return averageTime;
     }
 
-    public void setAverageTime(float averageTime) {
+    public void setAverageTime(double averageTime) {
         this.averageTime = averageTime;
     }
